test: synchronize (#273)
* refactor: rename & move synchro utils

* test: add tests on synchro

* test: improve tests
